ml_framework/resnet/requirements.txt
torch
pandas
torchvision
pytorch_lightning
sklearn
torchmetrics
pillow
